requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
pandas
numpy
pillow
gunicorn
dj-database-url
scikit-learn
whitenoise==6.7.0